+++ tinlok-core/src/commonMain/kotlin/tf/lotte/tinlok/net/tcp/TcpClientSocket.kt
@@ -23,13 +23,14 @@ import tf.lotte.tinlok.util.Unsafe
 public interface TcpClientSocket : TcpSocket, StreamingClientSocket<TcpConnectionInfo> {
     public companion object {
         /**
-         * Creates a new [TcpClientSocket] from a platform socket.
+         * Creates a new [TcpClientSocket] from a platform socket, using the specified [address]
+         * and [timeout].
          *
          * This method is unsafe as it can leak file descriptors.
          */
         @Unsafe
-        public fun unsafeOpen(address: TcpSocketAddress): TcpClientSocket {
-            return PlatformSockets.newTcpSynchronousSocket(address)
+        public fun unsafeOpen(address: TcpSocketAddress, timeout: Int): TcpClientSocket {
+            return PlatformSockets.newTcpSynchronousSocket(address, timeout)
         }
     }
 }